// src/api/openapi/models/NotificationResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationSettingsDataResponseDto } from './NotificationSettingsDataResponseDto';
import {
    NotificationSettingsDataResponseDtoFromJSON,
    NotificationSettingsDataResponseDtoFromJSONTyped,
    NotificationSettingsDataResponseDtoToJSON,
    NotificationSettingsDataResponseDtoToJSONTyped,
} from './NotificationSettingsDataResponseDto';
import type { NotificationsEnum } from './NotificationsEnum';
import {
    NotificationsEnumFromJSON,
    NotificationsEnumFromJSONTyped,
    NotificationsEnumToJSON,
    NotificationsEnumToJSONTyped,
} from './NotificationsEnum';

/**
 * 
 * @export
 * @interface NotificationResponseDto
 */
export interface NotificationResponseDto {
    /**
     * 
     * @type {boolean}
     * @memberof NotificationResponseDto
     */
    isEmailEnabled: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationResponseDto
     */
    isSmsEnabled: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationResponseDto
     */
    isAppPushEnabled: boolean;
    /**
     * 
     * @type {NotificationsEnum}
     * @memberof NotificationResponseDto
     */
    type: NotificationsEnum;
    /**
     * 
     * @type {NotificationSettingsDataResponseDto}
     * @memberof NotificationResponseDto
     */
    settingsJson: NotificationSettingsDataResponseDto;
}



/**
 * Check if a given object implements the NotificationResponseDto interface.
 */
export function instanceOfNotificationResponseDto(value: object): value is NotificationResponseDto {
    if (!('isEmailEnabled' in value) || value['isEmailEnabled'] === undefined) return false;
    if (!('isSmsEnabled' in value) || value['isSmsEnabled'] === undefined) return false;
    if (!('isAppPushEnabled' in value) || value['isAppPushEnabled'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('settingsJson' in value) || value['settingsJson'] === undefined) return false;
    return true;
}

export function NotificationResponseDtoFromJSON(json: any): NotificationResponseDto {
    return NotificationResponseDtoFromJSONTyped(json, false);
}

export function NotificationResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'isEmailEnabled': json['isEmailEnabled'],
        'isSmsEnabled': json['isSmsEnabled'],
        'isAppPushEnabled': json['isAppPushEnabled'],
        'type': NotificationsEnumFromJSON(json['type']),
        'settingsJson': NotificationSettingsDataResponseDtoFromJSON(json['settingsJson']),
    };
}

export function NotificationResponseDtoToJSON(json: any): NotificationResponseDto {
    return NotificationResponseDtoToJSONTyped(json, false);
}

export function NotificationResponseDtoToJSONTyped(value?: NotificationResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'isEmailEnabled': value['isEmailEnabled'],
        'isSmsEnabled': value['isSmsEnabled'],
        'isAppPushEnabled': value['isAppPushEnabled'],
        'type': NotificationsEnumToJSON(value['type']),
        'settingsJson': NotificationSettingsDataResponseDtoToJSON(value['settingsJson']),
    };
}