// src/api/openapi/models/NotificationSettingsRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationSettingsDataRequestDto } from './NotificationSettingsDataRequestDto';
import {
    NotificationSettingsDataRequestDtoFromJSON,
    NotificationSettingsDataRequestDtoFromJSONTyped,
    NotificationSettingsDataRequestDtoToJSON,
    NotificationSettingsDataRequestDtoToJSONTyped,
} from './NotificationSettingsDataRequestDto';
import type { NotificationsEnum } from './NotificationsEnum';
import {
    NotificationsEnumFromJSON,
    NotificationsEnumFromJSONTyped,
    NotificationsEnumToJSON,
    NotificationsEnumToJSONTyped,
} from './NotificationsEnum';

/**
 * 
 * @export
 * @interface NotificationSettingsRequestDto
 */
export interface NotificationSettingsRequestDto {
    /**
     * 
     * @type {boolean}
     * @memberof NotificationSettingsRequestDto
     */
    isEmailEnabled: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationSettingsRequestDto
     */
    isSmsEnabled: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationSettingsRequestDto
     */
    isAppPushEnabled: boolean;
    /**
     * 
     * @type {NotificationsEnum}
     * @memberof NotificationSettingsRequestDto
     */
    type: NotificationsEnum;
    /**
     * 
     * @type {NotificationSettingsDataRequestDto}
     * @memberof NotificationSettingsRequestDto
     */
    settingsJson: NotificationSettingsDataRequestDto;
}



/**
 * Check if a given object implements the NotificationSettingsRequestDto interface.
 */
export function instanceOfNotificationSettingsRequestDto(value: object): value is NotificationSettingsRequestDto {
    if (!('isEmailEnabled' in value) || value['isEmailEnabled'] === undefined) return false;
    if (!('isSmsEnabled' in value) || value['isSmsEnabled'] === undefined) return false;
    if (!('isAppPushEnabled' in value) || value['isAppPushEnabled'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('settingsJson' in value) || value['settingsJson'] === undefined) return false;
    return true;
}

export function NotificationSettingsRequestDtoFromJSON(json: any): NotificationSettingsRequestDto {
    return NotificationSettingsRequestDtoFromJSONTyped(json, false);
}

export function NotificationSettingsRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationSettingsRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'isEmailEnabled': json['isEmailEnabled'],
        'isSmsEnabled': json['isSmsEnabled'],
        'isAppPushEnabled': json['isAppPushEnabled'],
        'type': NotificationsEnumFromJSON(json['type']),
        'settingsJson': NotificationSettingsDataRequestDtoFromJSON(json['settingsJson']),
    };
}

export function NotificationSettingsRequestDtoToJSON(json: any): NotificationSettingsRequestDto {
    return NotificationSettingsRequestDtoToJSONTyped(json, false);
}

export function NotificationSettingsRequestDtoToJSONTyped(value?: NotificationSettingsRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'isEmailEnabled': value['isEmailEnabled'],
        'isSmsEnabled': value['isSmsEnabled'],
        'isAppPushEnabled': value['isAppPushEnabled'],
        'type': NotificationsEnumToJSON(value['type']),
        'settingsJson': NotificationSettingsDataRequestDtoToJSON(value['settingsJson']),
    };
}